% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkaeplot.R
\name{combine_pkaeplot}
\alias{combine_pkaeplot}
\title{Combine a list of figures from pkaeplot function}
\usage{
combine_pkaeplot(glist, rel_heights = c(3, 1), rel_widths = c(8, 2),
  legend = TRUE)
}
\arguments{
\item{glist}{list of figures generated with pkaeplot function}

\item{rel_heights}{numeric vector of relative columns heights of figures
for active and placebo arm (see cowplot::plot_grid for detail)}

\item{rel_widths}{numeric vector of relative columns widths of figures
and legend (see cowplot::plot_grid for detail)}

\item{legend}{whether to include legend in the plot}
}
\value{
A graphic object containing figures for active and placebo arm
}
\description{
\code{combine_pkaeplot} combines a list of figures from pkaeplot function
}
\details{
Instead of using this function, you can write your own function to combine
list of figures for greater flexibility
}
\examples{
combine_pkaeplot(glist)


}

