% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkaeplot.R
\name{pkaeplot}
\alias{pkaeplot}
\title{Overlay adverse event duration on concentration-time curve plot}
\usage{
pkaeplot(pk, ae, subj, ae.data.first.day = 0, scale.y.log10 = T,
  x.range = NULL, y.range = NULL, ae.col.var = "AE", ae.col.name = NULL,
  pk.col.var = NULL, ae.palette = c("#56B4E9", "#0072B2", "#D55E00"),
  ggtheme = NULL)
}
\arguments{
\item{pk}{dataset containing individual pharmacokinetic profiles}

\item{ae}{dataset containing adverse event profiles}

\item{ae.data.first.day}{assign 1 if the first study day is recorded as
DAY=1 in AE dataset and DAY=0 in PK dataset}

\item{scale.y.log10}{logical value controlling Y axis scale}

\item{x.range}{range of X axis to display}

\item{y.range}{range of Y axis to display for active arm}

\item{ae.col.var}{variable to set colors of AE (e.g. AE grade, AE type)}

\item{ae.col.name}{name appearing in legend for different colors of AE curves}

\item{pk.col.var}{variable to set colors of background PK curves (e.g. DOSE)}

\item{ae.palette}{color palette for AE scales}

\item{ggtheme}{specify ggplot theme (set NULL to use current theme)}
}
\value{
list of ggplot objects
}
\description{
\code{pkaeplot} creates a plot showing duration of adverse events in each subjects
overlayed on individual concentration-time curve
}
\details{
The function returns a list of ggplot objects, including (1) plot for active arm,
(2) plot for placebo arm, and (3) legend.
}
\examples{
pkaeplot(pk = pk.sim.profile.sparse,
         ae = ae.sim.profile,
         ae.data.first.day = 1,
         y.range=c(0.3,3))


}

